+++ src/main/java/com/crowdin/client/projectsgroups/model/Group.java
@@ -7,12 +7,12 @@ import java.util.Date;
 @Data
 public class Group {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
-    private Integer parentId;
-    private Integer organizationId;
-    private Integer userId;
+    private Long parentId;
+    private Long organizationId;
+    private Long userId;
     private Integer subgroupsCount;
     private Integer projectsCount;
     private Date createdAt;
