+++ src/main/java/com/crowdin/client/sourcefiles/SourceFilesApi.java
@@ -48,7 +48,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param offset    starting offset in the collection (default 0)
      * @return list of branches
      */
-    public ResponseList<Branch> listBranches(Integer projectId, String name, Integer limit, Integer offset) throws HttpException, HttpBadRequestException {
+    public ResponseList<Branch> listBranches(Long projectId, String name, Integer limit, Integer offset) throws HttpException, HttpBadRequestException {
         Map<String, Optional<Object>> queryParams = HttpConfig.buildUrlParams(
                 "name", Optional.ofNullable(name),
                 "limit", Optional.ofNullable(limit),
@@ -63,7 +63,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param request   request object
      * @return newly created branch
      */
-    public ResponseObject<Branch> addBranch(Integer projectId, AddBranchRequest request) throws HttpException, HttpBadRequestException {
+    public ResponseObject<Branch> addBranch(Long projectId, AddBranchRequest request) throws HttpException, HttpBadRequestException {
         BranchResponseObject branchResponseObject = this.httpClient.post(this.url + "/projects/" + projectId + "/branches", request, new HttpConfig(), BranchResponseObject.class);
         return ResponseObject.of(branchResponseObject.getData());
     }
@@ -73,7 +73,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param branchId  branch identifier
      * @return branch
      */
-    public ResponseObject<Branch> getBranch(Integer projectId, Integer branchId) throws HttpException, HttpBadRequestException {
+    public ResponseObject<Branch> getBranch(Long projectId, Long branchId) throws HttpException, HttpBadRequestException {
         BranchResponseObject branchResponseObject = this.httpClient.get(this.url + "/projects/" + projectId + "/branches/" + branchId, new HttpConfig(), BranchResponseObject.class);
         return ResponseObject.of(branchResponseObject.getData());
     }
@@ -82,7 +82,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param projectId project identifier
      * @param branchId  branch identifier
      */
-    public void deleteBranch(Integer projectId, Integer branchId) throws HttpException, HttpBadRequestException {
+    public void deleteBranch(Long projectId, Long branchId) throws HttpException, HttpBadRequestException {
         this.httpClient.delete(this.url + "/projects/" + projectId + "/branches/" + branchId, new HttpConfig(), Void.class);
     }
 
@@ -92,7 +92,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param request   request object
      * @return updated branch
      */
-    public ResponseObject<Branch> editBranch(Integer projectId, Integer branchId, List<PatchOperation> request) throws HttpException, HttpBadRequestException {
+    public ResponseObject<Branch> editBranch(Long projectId, Long branchId, List<PatchOperation> request) throws HttpException, HttpBadRequestException {
         BranchResponseObject groupResponseObject = this.httpClient.patch(this.url + "/projects/" + projectId + "/branches/" + branchId, request, new HttpConfig(), BranchResponseObject.class);
         return ResponseObject.of(groupResponseObject.getData());
     }
@@ -106,7 +106,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param offset      starting offset in the collection (default 0)
      * @return list of directories
      */
-    public ResponseList<Directory> listDirectories(Integer projectId, Integer branchId, Integer directoryId, Object recursion, Integer limit, Integer offset) throws HttpException, HttpBadRequestException {
+    public ResponseList<Directory> listDirectories(Long projectId, Long branchId, Long directoryId, Object recursion, Integer limit, Integer offset) throws HttpException, HttpBadRequestException {
         Map<String, Optional<Object>> queryParams = HttpConfig.buildUrlParams(
                 "branchId", Optional.ofNullable(branchId),
                 "directoryId", Optional.ofNullable(directoryId),
@@ -123,7 +123,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param request   request object
      * @return newly created directory
      */
-    public ResponseObject<Directory> addDirectory(Integer projectId, AddDirectoryRequest request) throws HttpException, HttpBadRequestException {
+    public ResponseObject<Directory> addDirectory(Long projectId, AddDirectoryRequest request) throws HttpException, HttpBadRequestException {
         DirectoryResponseObject post = this.httpClient.post(this.url + "/projects/" + projectId + "/directories", request, new HttpConfig(), DirectoryResponseObject.class);
         return ResponseObject.of(post.getData());
     }
@@ -133,7 +133,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param directoryId directory identifier
      * @return directory
      */
-    public ResponseObject<Directory> getDirectory(Integer projectId, Integer directoryId) throws HttpException, HttpBadRequestException {
+    public ResponseObject<Directory> getDirectory(Long projectId, Long directoryId) throws HttpException, HttpBadRequestException {
         DirectoryResponseObject directoryResponseObject = this.httpClient.get(this.url + "/projects/" + projectId + "/directories/" + directoryId, new HttpConfig(), DirectoryResponseObject.class);
         return ResponseObject.of(directoryResponseObject.getData());
     }
@@ -142,7 +142,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param projectId   project identifier
      * @param directoryId directory identifier
      */
-    public void deleteDirectory(Integer projectId, Integer directoryId) throws HttpException, HttpBadRequestException {
+    public void deleteDirectory(Long projectId, Long directoryId) throws HttpException, HttpBadRequestException {
         this.httpClient.delete(this.url + "/projects/" + projectId + "/directories/" + directoryId, new HttpConfig(), Void.class);
     }
 
@@ -152,7 +152,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param request     request object
      * @return updated directory
      */
-    public ResponseObject<Directory> editDirectory(Integer projectId, Integer directoryId, List<PatchOperation> request) throws HttpException, HttpBadRequestException {
+    public ResponseObject<Directory> editDirectory(Long projectId, Long directoryId, List<PatchOperation> request) throws HttpException, HttpBadRequestException {
         DirectoryResponseObject directoryResponseObject = this.httpClient.patch(this.url + "/projects/" + projectId + "/directories/" + directoryId, request, new HttpConfig(), DirectoryResponseObject.class);
         return ResponseObject.of(directoryResponseObject.getData());
     }
@@ -166,7 +166,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param offset      starting offset in the collection (default 0)
      * @return list of files
      */
-    public ResponseList<File> listFiles(Integer projectId, Integer branchId, Integer directoryId, Object recursion, Integer limit, Integer offset) throws HttpException, HttpBadRequestException {
+    public ResponseList<File> listFiles(Long projectId, Long branchId, Long directoryId, Object recursion, Integer limit, Integer offset) throws HttpException, HttpBadRequestException {
         Map<String, Optional<Object>> queryParams = HttpConfig.buildUrlParams(
                 "branchId", Optional.ofNullable(branchId),
                 "directoryId", Optional.ofNullable(directoryId),
@@ -183,7 +183,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param request   request object
      * @return newly created file
      */
-    public ResponseObject<File> addFile(Integer projectId, AddFileRequest request) throws HttpException, HttpBadRequestException {
+    public ResponseObject<File> addFile(Long projectId, AddFileRequest request) throws HttpException, HttpBadRequestException {
         FileResponseObject fileResponseObject = this.httpClient.post(this.url + "/projects/" + projectId + "/files", request, new HttpConfig(), FileResponseObject.class);
         return ResponseObject.of(fileResponseObject.getData());
     }
@@ -193,7 +193,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param fileId    file identifier
      * @return file
      */
-    public ResponseObject<File> getFile(Integer projectId, Integer fileId) throws HttpException, HttpBadRequestException {
+    public ResponseObject<File> getFile(Long projectId, Long fileId) throws HttpException, HttpBadRequestException {
         FileResponseObject fileResponseObject = this.httpClient.get(this.url + "/projects/" + projectId + "/files/" + fileId, new HttpConfig(), FileResponseObject.class);
         return ResponseObject.of(fileResponseObject.getData());
     }
@@ -204,7 +204,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param request   request object
      * @return updated file
      */
-    public ResponseObject<File> updateOrRestoreFile(Integer projectId, Integer fileId, UpdateOrRestoreFileRequest request) throws HttpException, HttpBadRequestException {
+    public ResponseObject<File> updateOrRestoreFile(Long projectId, Long fileId, UpdateOrRestoreFileRequest request) throws HttpException, HttpBadRequestException {
         FileResponseObject fileResponseObject = this.httpClient.put(this.url + "/projects/" + projectId + "/files/" + fileId, request, new HttpConfig(), FileResponseObject.class);
         return ResponseObject.of(fileResponseObject.getData());
     }
@@ -213,7 +213,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param projectId project identifier
      * @param fileId    file identifier
      */
-    public void deleteFile(Integer projectId, Integer fileId) throws HttpException, HttpBadRequestException {
+    public void deleteFile(Long projectId, Long fileId) throws HttpException, HttpBadRequestException {
         this.httpClient.delete(this.url + "/projects/" + projectId + "/files/" + fileId, new HttpConfig(), Void.class);
     }
 
@@ -223,7 +223,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param request   request object
      * @return updated file
      */
-    public ResponseObject<File> editFile(Integer projectId, Integer fileId, List<PatchOperation> request) throws HttpException, HttpBadRequestException {
+    public ResponseObject<File> editFile(Long projectId, Long fileId, List<PatchOperation> request) throws HttpException, HttpBadRequestException {
         FileResponseObject fileResponseObject = this.httpClient.patch(this.url + "/projects/" + projectId + "/files/" + fileId, request, new HttpConfig(), FileResponseObject.class);
         return ResponseObject.of(fileResponseObject.getData());
     }
@@ -233,7 +233,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param fileId    file identifier
      * @return file download link
      */
-    public ResponseObject<DownloadLink> downloadFile(Integer projectId, Integer fileId) throws HttpException, HttpBadRequestException {
+    public ResponseObject<DownloadLink> downloadFile(Long projectId, Long fileId) throws HttpException, HttpBadRequestException {
         DownloadLinkResponseObject downloadLinkResponseObject = this.httpClient.get(this.url + "/projects/" + projectId + "/files/" + fileId + "/download", new HttpConfig(), DownloadLinkResponseObject.class);
         return ResponseObject.of(downloadLinkResponseObject.getData());
     }
@@ -245,7 +245,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param offset    starting offset in the collection (default 0)
      * @return list of file revisions
      */
-    public ResponseList<FileRevision> listFileRevisions(Integer projectId, Integer fileId, Integer limit, Integer offset) throws HttpException, HttpBadRequestException {
+    public ResponseList<FileRevision> listFileRevisions(Long projectId, Long fileId, Integer limit, Integer offset) throws HttpException, HttpBadRequestException {
         Map<String, Optional<Object>> queryParams = HttpConfig.buildUrlParams(
                 "limit", Optional.ofNullable(limit),
                 "offset", Optional.ofNullable(offset)
@@ -260,7 +260,7 @@ public class SourceFilesApi extends CrowdinApi {
      * @param revisionId revision identifier
      * @return file revision
      */
-    public ResponseObject<FileRevision> getFileRevision(Integer projectId, Integer fileId, Integer revisionId) throws HttpException, HttpBadRequestException {
+    public ResponseObject<FileRevision> getFileRevision(Long projectId, Long fileId, Long revisionId) throws HttpException, HttpBadRequestException {
         FileRevisionResponseObject fileResponseObject = this.httpClient.get(this.url + "/projects/" + projectId + "/files/" + fileId + "/revisions/" + revisionId, new HttpConfig(), FileRevisionResponseObject.class);
         return ResponseObject.of(fileResponseObject.getData());
     }
