+++ src/main/java/com/crowdin/client/sourcefiles/model/File.java
@@ -8,14 +8,14 @@ import java.util.Date;
 @Data
 public class File {
 
-    private Integer id;
-    private Integer projectId;
-    private Integer branchId;
-    private Integer directoryId;
+    private Long id;
+    private Long projectId;
+    private Long branchId;
+    private Long directoryId;
     private String name;
     private String title;
     private Type type;
-    private Integer revisionId;
+    private Long revisionId;
     private String status;
     private Priority priority;
     private ImportOptions importOptions;
